Add input components
