+++ src/components/CalculatorForm.vue
@@ -1,71 +1,16 @@
 <template>
   <div>
     <div class="row">
-      <div class="col-12">
-        <div class="position-relative text-left">
-          <input
-            type="tel"
-            maxlength="3"
-            class="form-control"
-            v-model="formData.age.value"
-            @focus="onFocus(formData.age)"
-            @keypress="onTouch($event,formData.age)"
-            @blur="ageBlur(formData.age)"
-            autocomplete="off"
-          />
-          <label
-            class="input-label"
-            v-bind:class="{ active: formData.age.touched}"
-          >{{placeholderStr}}</label>
-          <span
-            id="heroCalculatorAge_error"
-            v-bind:class="{ 'validation-warning': formData.age.hasError}"
-          >{{formData.age.errorMessage}}</span>
-        </div>
-      </div>
+      <AgeInput :placeholderStr="placeholderStr" @onSetAge="setAge" ref="ageInput" />
     </div>
     <div class="row">
-      <div class="col-12">
-        <div class="position-relative text-left mt-4">
-          <input
-            type="tel"
-            class="form-control"
-            v-model="homeDisplay"
-            @keypress="onTouch($event,formData.homeValue)"
-            @blur="homeBlur(formData.homeValue)"
-            @focus="onFocus(formData.homeValue)"
-            required
-            autocomplete="off"
-          />
-          <label
-            class="input-label"
-            v-bind:class="{ active: formData.homeValue.touched}"
-          >Enter your home value</label>
-          <span
-            id="heroCalculatorHouseValue_error"
-            v-bind:class="{ 'validation-warning': formData.homeValue.hasError}"
-          >Minimum property value is $200,000</span>
-        </div>
-      </div>
+      <HomeValueInput @onSetHomeValue="setHomeValue" ref="homeInput" />
     </div>
+
     <div class="row mb-0">
       <div class="col-12 col-sm-8">
-        <div class="position-relative text-left mt-4 mb-5">
-          <input
-            type="tel"
-            maxlength="4"
-            class="form-control"
-            v-model="formData.postCode.value"
-            @keypress="onTouch($event,formData.postCode)"
-            required
-          />
-          <label
-            class="input-label"
-            v-bind:class="{ active: formData.postCode.touched}"
-          >Enter your postcode</label>
-        </div>
+        <PostcodeInput @onSetPostcode="setPostCode" ref="postcodeInput" />
       </div>
-
       <div class="mx-auto my-4">
         <div
           class="btn btn-lg btn-primary small-calculator-calculate-btn js-hero-calculator-calculate-btn my-0"
@@ -81,6 +26,9 @@
 
 <script>
 import { strip } from "../util/common";
+import AgeInput from "./AgeInput.vue";
+import HomeValueInput from "./HomeValueInput.vue";
+import PostcodeInput from "./PostcodeInput.vue";
 import { mapGetters } from "vuex";
 export default {
   name: "Calculator-form",
@@ -90,31 +38,15 @@ export default {
       required: true
     }
   },
+  components: {
+    AgeInput,
+    HomeValueInput,
+    PostcodeInput
+  },
 
-  watch: {
+  watch: {//notify update
     getUpdateResetForm() {
-      this. buttonText= "Calculate";
-      this.formData.age = {
-        value: undefined,
-        touched: false,
-        hasError: false,
-        errorMessage: "Minimum age is 60 years",
-        previousValue: undefined
-      };
-      this.formData.homeValue = {
-        value: undefined,
-        touched: false,
-        hasError: false,
-        errorMessage: "Minimum property value is $200,000",
-        previousValue: undefined
-      };
-      this.formData.postCode = {
-        value: undefined,
-        touched: false,
-        hasError: false,
-        errorMessage: "",
-        previousValue: undefined
-      };
+      this.buttonText = "Calculate";
     }
   },
   data() {
@@ -125,104 +57,42 @@ export default {
           value: undefined,
           touched: false,
           hasError: false,
-          errorMessage: "Minimum age is 60 years",
           previousValue: undefined
         },
         homeValue: {
           value: undefined,
           touched: false,
           hasError: false,
-          errorMessage: "Minimum property value is $200,000",
           previousValue: undefined
         },
-        postCode: {
+        postcode: {
           value: undefined,
           touched: false,
           hasError: false,
-          errorMessage: "",
           previousValue: undefined
         }
       }
     };
   },
   computed: {
-     ...mapGetters("ui", ["getUpdateResetForm"]),
-    ageValue() {
-      return this.formData.age.value;
-    },
-    homeValue() {
-      return this.formData.homeValue.value;
-    },
-    homeDisplay: {
-      get() {
-        if (this.formData.homeValue.value) {
-          const stripped = strip(this.formData.homeValue.value);
-          if (stripped) {
-            const formattedStripped = parseInt(stripped).toLocaleString("en", {
-              currency: "USD",
-              minimumFractionDigits: 0,
-              maximumFractionDigits: 7
-            });
-            return "$" + formattedStripped;
-          } else {
-            return "$";
-          }
-        }
-
-        return "";
-      },
-      set(newValue) {
-        const strippedZero = newValue.replace(/0/g, "");
-        if (strippedZero === "$") {
-          this.formData.homeValue.value = 0;
-        }
-        this.formData.homeValue.value = newValue;
-      }
-    }
+    ...mapGetters("ui", ["getUpdateResetForm"])
   },
 
   methods: {
-    onTouch(e, input) {
-      input.touched = true;
-      if (e.charCode >= 48 && e.charCode <= 57) {
-        return true;
-      }
-      e.preventDefault();
+    setAge(age) {
+      this.formData.age = age;
     },
-    onFocus(input) {
-      input.previousValue = input.value;
-      input.value = "";
+    setHomeValue(homeValue) {
+      this.formData.homeValue = homeValue;
     },
-    ageBlur(input) {
-      if (!input.value) {
-        input.value = input.previousValue ? input.previousValue : 0;
-        input.touched = true;
-      }
-      if (!this.ageValidate(input.value)) {
-        input.hasError = true;
-      } else {
-        input.hasError = false;
-      }
-      input.errorMessage = "Minimum age is 60 years";
-      if (input.value > 120) {
-        input.errorMessage = "Are you intending to beat the world record? :P";
-      }
+    setPostCode(postcode) {
+      this.formData.postcode = postcode;
     },
-    homeBlur(input) {
-      input.touched = true;
-      if (!input.value) {
-        input.value = input.previousValue ? input.previousValue : "$";
-      }
-      if (!this.homeValueValidate(input.value)) {
-        input.hasError = true;
-      } else {
-        input.hasError = false;
-      }
-    },
+
     async calculate() {
       if (this.validate()) {
         await this.$store.dispatch("postcode/getPostcodeValidation", {
-          postcode: this.formData.postCode.value
+          postcode: this.formData.postcode.value
         });
         const lvr = await this.$store.dispatch("age/getLvrByAge", {
           age: this.formData.age.value
@@ -240,45 +110,30 @@ export default {
     },
     validate() {
       let valid = true;
+      //trigger children component get data
+      this.$refs.ageInput.onBlur();
+      this.$refs.homeInput.onBlur();
+      this.$refs.postcodeInput.onBlur();
       if (
-        !this.formData.age.value ||
-        !this.ageValidate(this.formData.age.value)
+        !this.formData.age.value
       ) {
-        this.formData.age.hasError = true;
-        this.formData.age.touched = true;
-        if (!this.formData.age.value) {
-          this.formData.age.value = 0;
-        }
+      
         valid = false;
       }
       if (
-        !this.formData.homeValue.value ||
-        !this.homeValueValidate(this.formData.homeValue.value)
+        !this.formData.homeValue.value
       ) {
-        this.formData.homeValue.hasError = true;
-        this.formData.homeValue.touched = true;
-        if (!this.formData.homeValue.value) {
-          this.formData.homeValue.value = "$";
-        }
         valid = false;
       }
       if (!this.postCodeValidate()) {
         valid = false;
       }
-
       return valid;
-    },
-    ageValidate(value) {
-      return value >= 60 && value <= 120;
-    },
-    homeValueValidate(value) {
-      var stripped = strip(value);
-      return stripped && stripped >= 200000;
     },
     postCodeValidate() {
       return (
-        this.formData.postCode.value &&
-        this.formData.postCode.value.length === 4
+        this.formData.postcode.value &&
+        this.formData.postcode.value.length === 4
       );
     }
   }
